Recode - recoded most of the source - removed bloat includes - once again improved render - finished extra windows addons - fixed crash on unloading - completely rewrote plot's logic, making it custom - completely fixed world modulation / hooking directly instead of looping and hooking isUsingDebugProops to directly remove static props issues - temporiary removed chams, because old solution was "just working"

+++ SDK/Color.hpp
@@ -3,7 +3,7 @@
 #include <array>
 #include <stdexcept>
 
-using ImU32 = unsigned int;
+using ImU32 = size_t;
 struct ImVec4;
 
 // 0.0 - 1.0, you can pass ints as 0-255 though
